requirements.txt
fastapi
uvicorn
torch
numpy
soundfile
librosa
python-multipart
python-dotenv
